Extract navigation into router file
