+++ src/App.tsx
@@ -1,72 +1,12 @@
 import * as React from 'react';
-import { Text, View, StyleSheet, Dimensions } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { MaterialCommunityIcons } from '@expo/vector-icons';
+import { NavTabs } from './Router';
 
-import { MapScreen } from './screens/Map/MapScreen';
-import { ProfileScreen } from './screens/Profile/ProfileScreen';
-import { FriendsScreen } from './screens/Friends/FriendsScreen';
-import { FavouritesScreen } from './screens/Favourites/FavouritesScreen';
-
-const Tab = createBottomTabNavigator();
-
-function MyTabs() {
-  return (
-    <Tab.Navigator
-      initialRouteName="Map"
-      screenOptions={{
-        tabBarActiveTintColor: '#219f94',
-      }}
-    >
-      <Tab.Screen
-        name="Map"
-        component={MapScreen}
-        options={{
-          tabBarLabel: 'Explore',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons name="map" color={color} size={size} />
-          ),
-        }}
-      />
-      <Tab.Screen
-        name="Favourites"
-        component={FavouritesScreen}
-        options={{
-          tabBarLabel: 'Favourites',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons name="heart" color={color} size={size} />
-          ),
-        }}
-      />
-      <Tab.Screen
-        name="Friends"
-        component={FriendsScreen}
-        options={{
-          tabBarLabel: 'Friends',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons name="account-multiple-outline" color={color} size={size} />
-          ),
-        }}
-      />
-      <Tab.Screen
-        name="Profile"
-        component={ProfileScreen}
-        options={{
-          tabBarLabel: 'Profile',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons name="account" color={color} size={size} />
-          ),
-        }}
-      />
-    </Tab.Navigator>
-  );
-}
 
 export default function App() {
   return (
     <NavigationContainer>
-      <MyTabs />
+      <NavTabs/>
     </NavigationContainer>
   );
 }
